refactor: Change directory name
