+++ src/components/admin/party/ApplicationList/PartyRowDetail.tsx
